halt fine-uploader upload until after image loaded
The issue: some jpegs are taken in one orientation and then rotated
on a device and exported such that the display of the image doesn't
seem to match the dimensions. The result is you upload what you
think is a portrait image but is actually a landscape image, which
EXIF data set to rotate the image. This then gives un-expected results
in the image cropping tool and on the front end.

The fix: halt the fine-uploader by using a promise on the submit
callback of fine-uploader, until after the image has loaded and
JavaScript has had a change to grab the image dimensions.

(previously the upload was completing before the image had been loaded
and so image dimensions were not being sent)