Prevent HasSlug::getUtf8Slug() from guessing encoding by default
By default, the method was guessing from a list of all possible encodings what the encoding of the given string was. This change allows the method caller to supply the list of possible encodings. If no encodings are provided, it will use the `default_charset` setting. See https://www.php.net/manual/en/function.mb-convert-encoding.php.